defined variables

+++ 0x00-python_variable_annotations/4-define_variables.py
@@ -3,7 +3,7 @@
   This module defines and annotates specified values
 """
 
-a: float = 1
+a: int = 1
 pi: float = 3.14
 i_understand_annotations: bool = True
 school: str = 'Holberton'
